Add basics for bots supports (#78)
Add the basics for bot supports in lobbies.

I chose to use a different field for bots instead of cramming them
inside the `players` object. Makes the client code a little bit more
complex to display the rooms, but I think it'll work better when we
evolve players and bots, without running into the same problem of json
merge patch + sum type.

I also added a concept of ID for bots, that the server generates because
otherwise it's not clear how to identify a bot, and we need an ID for
remove & lobby/updated